+++ native/src/VibeSnake.Rules/Properties/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("VibeSnake.Rules.Tests")]
+[assembly: InternalsVisibleTo("VibeSnake.Game")]